Create libuv transport
This transport is largely modeled after the existing TCP transport.
There is no support for the old style and only for the new style
algorithms (there are no createSendBuffer/createRecvBuffer functions).

This commit adds the basic functionality but more work is needed to
properly deal with errors and timeouts.

It requires libuv to be available. Either it should already be defined
as a build target, if included by another project's build, or it must
be installed and discoverable through pkg-config.

ghstack-source-id: ad26fce87403621ee0e30df77dab2c705382bc2f
Pull Request resolved: https://github.com/facebookincubator/gloo/pull/212